refactor segment components
