+++ frontend/src/component/common/ConstraintsList/StrategyEvaluationItem/StrategyEvaluationItem.tsx
@@ -12,6 +12,7 @@ const StyledContainer = styled('div')(({ theme }) => ({
     gap: theme.spacing(1),
     alignItems: 'center',
     fontSize: theme.typography.body2.fontSize,
+    padding: theme.spacing(2, 3),
 }));
 
 const StyledType = styled('span')(({ theme }) => ({
