+++ frontend/src/component/common/SegmentItem/SegmentItem.tsx
@@ -1,6 +1,5 @@
-import { useState, type VFC } from 'react';
+import { useMemo, useState, type FC } from 'react';
 import { Link } from 'react-router-dom';
-import DonutLarge from '@mui/icons-material/DonutLarge';
 import type { ISegment } from 'interfaces/segment';
 import {
     Accordion,
@@ -10,132 +9,106 @@ import {
     styled,
     Typography,
 } from '@mui/material';
-import { ConstraintAccordionList } from 'component/common/ConstraintAccordion/ConstraintAccordionList/ConstraintAccordionList';
-import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
+import { StrategyEvaluationItem } from 'component/common/ConstraintsList/StrategyEvaluationItem/StrategyEvaluationItem';
+import { ConstraintItem } from 'component/common/ConstraintsList/ConstraintItem/ConstraintItem';
+import { objectId } from 'utils/objectId';
+import { ConstraintsList } from 'component/common/ConstraintsList/ConstraintsList';
 
-interface ISegmentItemProps {
+type SegmentItemProps = {
     segment: Partial<ISegment>;
     isExpanded?: boolean;
     disabled?: boolean | null;
     constraintList?: JSX.Element;
     headerContent?: JSX.Element;
-}
+};
 
-const StyledAccordion = styled(Accordion, {
-    shouldForwardProp: (prop) => prop !== 'isDisabled',
-})<{ isDisabled: boolean | null }>(({ theme, isDisabled }) => ({
-    border: `1px solid ${theme.palette.divider}`,
-    '&.segment-accordion': {
-        borderRadius: theme.shape.borderRadiusMedium,
-    },
+const StyledAccordion = styled(Accordion)(({ theme }) => ({
     boxShadow: 'none',
     margin: 0,
-    transition: 'all 0.1s ease',
-    '&:before': {
-        opacity: '0 !important',
-    },
-    '&.Mui-expanded': { backgroundColor: theme.palette.neutral.light },
-    backgroundColor: isDisabled
-        ? theme.palette.envAccordion.disabled
-        : theme.palette.background.paper,
+    padding: 0,
 }));
 
 const StyledAccordionSummary = styled(AccordionSummary)(({ theme }) => ({
-    margin: theme.spacing(0, 0.5),
+    padding: 0,
     fontSize: theme.typography.body2.fontSize,
-    '.MuiAccordionSummary-content': {
-        display: 'flex',
-        alignItems: 'center',
+    '.MuiAccordionSummary-content, .Mui-expanded.MuiAccordionSummary-content': {
+        margin: 0,
     },
 }));
 
-const StyledLink = styled(Link)(({ theme }) => ({
+const StyledAccordionDetails = styled(AccordionDetails)(({ theme }) => ({
+    padding: theme.spacing(0, 2, 3),
+}));
+
+const StyledLink = styled(Link)({
     textDecoration: 'none',
-    marginLeft: theme.spacing(1),
     '&:hover': {
         textDecoration: 'underline',
     },
+});
+
+const StyledActionsContainer = styled('div')(({ theme }) => ({
+    display: 'flex',
+    alignItems: 'center',
+    marginLeft: 'auto',
+    marginRight: theme.spacing(2),
 }));
 
-const StyledText = styled('span', {
-    shouldForwardProp: (prop) => prop !== 'disabled',
-})<{ disabled: boolean | null }>(({ theme, disabled }) => ({
-    color: disabled ? theme.palette.text.secondary : 'inherit',
+const StyledButton = styled(Button)(({ theme }) => ({
+    fontSize: theme.typography.body2.fontSize,
 }));
 
-export const SegmentItem: VFC<ISegmentItemProps> = ({
+export const SegmentItem: FC<SegmentItemProps> = ({
     segment,
     isExpanded,
     headerContent,
     constraintList,
-    disabled = false,
 }) => {
     const [isOpen, setIsOpen] = useState(isExpanded || false);
 
+    const constraints = useMemo(() => {
+        if (constraintList) {
+            return constraintList;
+        }
+
+        if (segment.constraints?.length) {
+            return (
+                <ConstraintsList>
+                    {segment.constraints.map((constraint, index) => (
+                        <ConstraintItem
+                            key={`${objectId(constraint)}-${index}`}
+                            {...constraint}
+                        />
+                    ))}
+                </ConstraintsList>
+            );
+        }
+
+        return <Typography>This segment has no constraints.</Typography>;
+    }, [constraintList, segment.constraints]);
+
     return (
-        <StyledAccordion
-            className='segment-accordion'
-            isDisabled={disabled}
-            expanded={isOpen}
-        >
+        <StyledAccordion expanded={isOpen}>
             <StyledAccordionSummary id={`segment-accordion-${segment.id}`}>
-                <DonutLarge
-                    sx={(theme) => ({
-                        mr: 1,
-                        color: disabled
-                            ? theme.palette.neutral.border
-                            : theme.palette.secondary.main,
-                    })}
-                />
-                <StyledText disabled={disabled}>Segment:</StyledText>
-                <StyledLink to={`/segments/edit/${segment.id}`}>
-                    {segment.name}
-                </StyledLink>
-                <ConditionallyRender
-                    condition={Boolean(headerContent)}
-                    show={headerContent}
-                />
-                <ConditionallyRender
-                    condition={!isExpanded}
-                    show={
-                        <Button
+                <StrategyEvaluationItem type='Segment'>
+                    <StyledLink to={`/segments/edit/${segment.id}`}>
+                        {segment.name}
+                    </StyledLink>
+                </StrategyEvaluationItem>
+                {headerContent ? headerContent : null}
+                {!isExpanded ? (
+                    <StyledActionsContainer>
+                        <StyledButton
                             size='small'
                             variant='outlined'
                             onClick={() => setIsOpen((value) => !value)}
-                            sx={{
-                                my: 0,
-                                ml: 'auto',
-                                fontSize: (theme) =>
-                                    theme.typography.body2.fontSize,
-                            }}
                         >
                             {isOpen ? 'Close preview' : 'Preview'}
-                        </Button>
-                    }
-                />
+                        </StyledButton>
+                    </StyledActionsContainer>
+                ) : null}
             </StyledAccordionSummary>
-            <AccordionDetails sx={{ pt: 0 }}>
-                <ConditionallyRender
-                    condition={Boolean(constraintList)}
-                    show={constraintList}
-                    elseShow={
-                        <ConditionallyRender
-                            condition={(segment?.constraints?.length || 0) > 0}
-                            show={
-                                <ConstraintAccordionList
-                                    constraints={segment!.constraints!}
-                                    showLabel={false}
-                                />
-                            }
-                            elseShow={
-                                <Typography>
-                                    This segment has no constraints.
-                                </Typography>
-                            }
-                        />
-                    }
-                />
-            </AccordionDetails>
+            <StyledAccordionDetails>{constraints}</StyledAccordionDetails>
         </StyledAccordion>
     );
 };
